requirements.txt
Django
python-decouple
django-debug-toolbar

